requirements.txt
requests
beautifulsoup4
readability-lxml
pyjwt
sqlalchemy
schedule
pytrends
python-dotenv